+++ Form_NSST/Form_Empty/Image.h
@@ -2,7 +2,7 @@
 #include <Windows.h>
 
 BYTE* LoadBMP(int% width, int% height, long% size, LPCTSTR bmpfile);
-float* ConvertBMPToIntensity(BYTE* Buffer, int width, int height);
+BYTE* ConvertBMPToIntensity(BYTE* Buffer, int width, int height);
 BYTE* ConvertIntensityToBMP(BYTE* Buffer, int width, int height, long% newsize);
 bool SaveBMP(BYTE* Buffer, int width, int height, long paddedsize, LPCTSTR bmpfile);
 float* ConvertBMPToYCbCr(BYTE* Buffer, int width, int height);
